// src/models/Scalars.ts
export type Scalars = {
    active_carriers: number;
    active_clients: number;
    average_margin_perc: number;
    avg_order_margin_abs: number;
    avg_order_revenue: number;
    new_carriers: number;
    new_clients: number;
    total_assigned_count: number;
    total_margin_abs: number;
    total_order_count: number;
    total_revenue: number;
  };